+++ ChangeLog.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-- The client now uses the V2 GRPC API exposed by the node. This introduces
-  some potentially breaking changes:
+- The client now uses the V2 GRPC API exposed by the node. This introduces some breaking
+  changes:
+    - The node serves the V2 GRPC API at port 20000 by default, and the default value of
+      the `--grpc-port` option has been updated to reflect this.
     - `raw` family of commands:
         - Commands `raw GetTransactionStatusInBlock`, `raw StartBaker`, `raw StopBaker`,
           `raw JoinNetwork` and ` raw LeaveNetwork` have been removed.
@@ -26,6 +28,7 @@
         - Slight changes to `raw GetPeerData` output. Notably the catch-up status and
           consensus-related details about the peer is elaborated upon.
         - `raw SendTransaction` no longer allows for specifying a network ID.
+    - `consensus show-parameters` now additionally prints the election difficulty.
     - Slight changes to error message information and their phrasing for other families
       of commands.